+++ C++/Recursion/TowerOfHanoi/TowerOfHanoi.cpp
@@ -31,3 +31,20 @@ int main()
 
     return 0;
 }
+
+/*Explanation:
+
+The towerOfHanoi function is defined with the same logic as in the C code.
+
+The base case and the recursive steps are the same as explained before.
+
+The main function is also similar to the C code, but it uses std::cout and std::cin for input/output instead of printf and scanf.
+
+The user is prompted to enter the number of disks they want to solve the Tower of Hanoi puzzle for.
+
+The towerOfHanoi function is called with the user-inputted number of disks, the source peg 'A', the destination peg 'C', and the auxiliary peg 'B'.
+
+The function executes and prints each move required to solve the puzzle using std::cout.
+
+Finally, the main function returns 0, indicating successful execution of the program.
+*/
