+++ C/Recursion/TowerOfHanoi/TowerOfHanoi.c
@@ -1,4 +1,38 @@
-Explanation:
+#include <stdio.h>
+
+// Function to move a disk from the source peg to the destination peg
+void towerOfHanoi(int n, char source, char destination, char auxiliary)
+{
+    // Base case: If only one disk is present, move it directly from source to destination
+    if (n == 1)
+    {
+        printf("Move disk 1 from peg %c to peg %c\n", source, destination);
+        return;
+    }
+
+    // Move (n-1) disks from the source peg to the auxiliary peg
+    towerOfHanoi(n - 1, source, auxiliary, destination);
+
+    // Move the remaining disk from the source peg to the destination peg
+    printf("Move disk %d from peg %c to peg %c\n", n, source, destination);
+
+    // Move the (n-1) disks from the auxiliary peg to the destination peg
+    towerOfHanoi(n - 1, auxiliary, destination, source);
+}
+
+int main()
+{
+    int numDisks;
+    printf("Enter the number of disks: ");
+    scanf("%d", &numDisks);
+
+    printf("Tower of Hanoi solution:\n");
+    towerOfHanoi(numDisks, 'A', 'C', 'B');
+
+    return 0;
+}
+
+/*Explanation:
 
 The towerOfHanoi function is a recursive function that takes four parameters:
 
@@ -20,3 +54,4 @@ The towerOfHanoi function is then called with the user-inputted number of disks,
 The function executes and prints each move required to solve the puzzle.
 
 Finally, the main function returns 0, indicating successful execution of the program.
+*/
